% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{plot.area}
\alias{plot.area}
\title{Plotting all the points generated inside the constraints}
\usage{
\method{plot}{area}(x)
}
\value{
It basically return a graph with all the points inside the boundaries
}
\description{
Provides a graph with all the points generated inside the boundaries. It makes senses only when the output of a function
has an object named as "points". We are basically subsetting the output of estimate_area() with points as one of the output of
estimate_area() is called "points"
}
\examples{
plot.area(estimate_area(B = 500, seed = 25))
}
\author{
Özgür Aydemir, Sophie La Gennusa, Louis del Perugia, Daniel Szenes, Francesca Darino
}
