% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area.R
\name{point_constraint}
\alias{point_constraint}
\title{Define the contstraint of points}
\usage{
point_constraint(x)
}
\value{
True or False
}
\description{
This function is only necessary for the function estimate_area() because it defines the points
that are inside of the boundaries and than will be used to calculate the area of the shape
}
\author{
Özgür Aydemir, Sophie La Gennusa, Louis del Perugia, Daniel Szenes, Francesca Darino
}
